# lib/api-zod/src/generated/types/downloadResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DownloadResponseStatus } from "./downloadResponseStatus";

export interface DownloadResponse {
  status: DownloadResponseStatus;
  title?: string | null;
  thumbnail?: string | null;
  video_url: string;
  quality: string;
  author?: string | null;
  duration?: number | null;
  cached?: boolean;
}