# lib/api-zod/src/generated/types/errorResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ErrorResponseStatus } from "./errorResponseStatus";

export interface ErrorResponse {
  status: ErrorResponseStatus;
  message: string;
}